Fix early object-creation bug.
In objects, we were storing singleton instances, which were created
when an object was registered at static initialization time. This
meant that the constructor of an object was called immediately when
it was registered, even if the program intended to mock out that
object.

To fix this, we store factory functions capable of creating an object,
and lazily instantiate the singletons when asked (at which point we
know whether we want the mock implementation or the real one).