Resolve cross-library symbol visibility issue.
When test.cpp.o links with libself-driving-sim.a as a static library,
it only pulls in symbols that are explicitly referenced in the test
object. Since all of the static initialization code in the library
is not explicitly referenced (rather, it is referenced indirectly
through objects::get, etc), these symbols are not brought in, and
no objects are registered at static initialization time.

To fix this, we build the library as a shared library
libself-driving-sim.so so that all the symbols, including the static
registration structs, are brought in at runtime.